# openapi.yaml
openapi: 3.0.3
info:
  title: CampusHub Reservation API
  version: 1.0.0
  description: Lab 2 OpenAPI contract for CampusHub resources and reservations.
servers:
  - url: http://localhost:3000/api/v1
paths:
  /resources:
    get:
      summary: List campus resources
      operationId: listResources
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum:
              - ROOM
              - EQUIPMENT
              - LAB
      responses:
        "200":
          description: Resources retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resource"

  /reservations:
    post:
      summary: Create a reservation
      operationId: createReservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Reservation"
      responses:
        "201":
          description: Reservation created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "409":
          description: Resource is already reserved during the requested time block
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /reservations/user/{userId}:
    get:
      summary: Get active reservations for a user
      operationId: getReservationsByUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Active reservations for the user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reservation"

components:
  schemas:
    Resource:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - type
        - isAvailable
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum:
            - ROOM
            - EQUIPMENT
            - LAB
        isAvailable:
          type: boolean

    Reservation:
      type: object
      additionalProperties: false
      required:
        - id
        - resourceId
        - userId
        - startTime
        - endTime
        - status
      properties:
        id:
          type: string
        resourceId:
          type: string
        userId:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        status:
          type: string
          enum:
            - PENDING
            - CONFIRMED
            - CANCELLED

    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string